Reorganize user login code
