+++ client/src/components/NavBar.vue
@@ -1,12 +1,10 @@
 <script setup lang="ts">
 import { RouterLink } from "vue-router";
 import { ref } from "vue";
-import { type User, getProducts, getUsers } from "../model/users";
-
+import LoginBadge from './LoginBadge.vue';
 import { getSession, login } from '@/model/session'
 
 let isActive = ref(false);
-let isActiveLogin = ref(false);
 
 function toggleMenu() {
   isActive.value = !isActive.value;
@@ -15,12 +13,6 @@ function toggleMenu() {
 function toggleMenuLogin() {
   isActiveLogin.value = !isActiveLogin.value;
 }
-
-const session = getSession()
-
-const users = ref([] as User[]);
-
-users.value = getUsers();
 </script>
 
 <template>
@@ -42,40 +34,7 @@ users.value = getUsers();
 
         <RouterLink to="friends" class="navbar-item"> Friends </RouterLink>
       </div>
-      <div class="navbar-end">
-        <div class="navbar-item">
-          <div v-if="session.user != null" class="navbar-item">
-            {{ session.user.name }}
-            <img :src="session.user.profilePicture"/>
-          </div>
-          <div class="buttons">
-            <a v-if="session.user == null" class="button is-black">
-              <strong>Sign up</strong>
-            </a>
-            <a v-if="session.user != null" class="button is-black" @click.prevent="session.user = user">
-              <strong>Logout</strong>
-            </a>
-            <div @click="toggleMenuLogin" :class="{ 'dropdown is-active': isActiveLogin }">
-              <div v-if="session.user == null" class="dropdown-trigger">
-                <button class="button" aria-haspopup="true" aria-controls="dropdown-menu">
-                  <span> Login </span>
-                  <span class="icon is-small">
-                    <i class="fas fa-angle-down" aria-hidden="true"></i>
-                  </span>
-                </button>
-              </div>
-              <div class="dropdown-menu" id="dropdown-menu" role="menu">
-                <div class="dropdown-content">
-                  <div v-for="user in users" :key="user.id">
-                    <a href="#" @click.prevent="session.user = user" class="dropdown-item">
-                      {{ user.name }}</a>
-                  </div>
-                </div>
-              </div>
-            </div>
-          </div>
-        </div>
-      </div>
+      <LoginBadge />
     </div>
   </nav>
 </template>
